Bunch of uncommitted changes.
Switch to using BSP instead of HintedTile as the main tiled
layout. Drop the "full with borders and struts" layout (never used
it), keep just BSP and proper fullscreen. Add keybindings for BSP.

Add a module for switching to xkb sublayouts, and a keybindings for
the first and second layouts.

Update Makefile and switch to make for rebuilding (from mod-x), in
order to pass -lxklavier. This is needed for the xkb layout switching.